+++ indico/modules/events/editing/operations.py
@@ -158,7 +158,7 @@ def review_editable_revision(revision, editor, action, comment, tags, files=None
                                        tags=tags)
         revision.editable.revisions.append(new_revision)
     db.session.flush()
-    notify_editor_judgment(revision, editor)
+    notify_editor_judgment(new_revision, revision.user, action)
     logger.info('Revision %r reviewed by %s [%s]', revision, editor, action.name)
     _log_review(revision.editable, LogKind.positive, f'Revision for {revision.editable.log_title} reviewed',
                 old_state=old_state)
@@ -183,7 +183,7 @@ def confirm_editable_changes(revision, submitter, action, comment):
     if action == EditingConfirmationAction.accept:
         _ensure_publishable_files(revision)
     db.session.flush()
-    notify_submitter_confirmation(revision, submitter, action)
+    notify_submitter_confirmation(new_revision, submitter, action)
     logger.info('Revision %r confirmed by %s [%s]', revision, submitter, action.name)
     _log_review(revision.editable, LogKind.positive, f'Revision for {revision.editable.log_title} confirmed',
                 old_state=old_state)