+++ CHANGES.rst
@@ -63,6 +63,7 @@ Improvements
 - Do not bother people registering using an invitation link with a CAPTCHA (:pr:`6095`)
 - Add option to allow people to register using an invitation link even if the event is
   restricted (:pr:`6094`)
+- Improve editing notifications emails (:issue:`6027`, :pr:`6042`)
 
 Bugfixes
 ^^^^^^^^